requirements.txt
aiogram==3.24.0
SQLAlchemy>=2.0.36,<3.0
greenlet>=3.0.3
aiosqlite==0.19.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.3

